Refactor TextLineDatasetOp
